Added: admin credentials

+++ SetNoti.py
@@ -1,5 +1,10 @@
 import os
+# from posixpath import split
 import sys
+from BOOT import listMails
+from Mails.SendPDF import *
+from Mails.BootMails import *
+
 
 def FunList(List):
     """[Takes list and prints it vertically assigning each item its index number on CLI]
@@ -41,9 +46,25 @@ def AskNoti():
     """[To replace the notification from one folder to the storage and send the email to the respective members.]
     """
     C_Noti = input("Enter the path of the notification: ")
+    if not os.path.exists(C_Noti):
+        print("Path doesnot exists...")
+        return
     T_Path = os.path.join(GetPath(), os.path.split(C_Noti)[1])
+    if not "Student" in T_Path:
+        sendPDFMail("A new notification has been added.", C_Noti, subject="A new notification", recievers=listMails)
+    else:
+        with open(os.path.join(T_Path, "CREs.txt"), "r") as f:
+            R = DC(f.read()).split("\n")[2].split(": ")[1].strip()
+        sendPDFMail("A new notification has been added.", C_Noti, subject="A new notification", recievers=[R])
     os.replace(C_Noti, T_Path)
     # ! sendMailNoti(T_path)
-
+def AdminLog():
+    Pass = PasswordGenerator()
+    SendBootCre(Pass, Topic="admin")
+    if input("Enter the password sent you through mail: ") == Pass:
+        AskNoti()
+    else:
+        print("Wrong Credentials...")
+    
 if __name__ == '__main__':
     AskNoti()
